+++ packages/instabug_flutter/android/src/main/java/com/instabug/flutter/modules/InstabugApi.java
@@ -6,13 +6,16 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.util.Log;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
+
 import com.instabug.flutter.generated.InstabugPigeon;
 import com.instabug.flutter.util.ArgsRegistry;
 import com.instabug.flutter.util.Reflection;
 import com.instabug.flutter.util.ThreadManager;
+import com.instabug.flutter.util.privateViews.ScreenshotCaptor;
 import com.instabug.library.Feature;
 import com.instabug.library.Instabug;
 import com.instabug.library.InstabugColorTheme;
@@ -21,13 +24,17 @@ import com.instabug.library.IssueType;
 import com.instabug.library.Platform;
 import com.instabug.library.ReproConfigurations;
 import com.instabug.library.featuresflags.model.IBGFeatureFlag;
+import com.instabug.library.internal.crossplatform.InternalCore;
 import com.instabug.library.internal.module.InstabugLocale;
 import com.instabug.library.invocation.InstabugInvocationEvent;
 import com.instabug.library.model.NetworkLog;
+import com.instabug.library.screenshot.instacapture.ScreenshotRequest;
 import com.instabug.library.ui.onboarding.WelcomeMessage;
+
 import io.flutter.FlutterInjector;
 import io.flutter.embedding.engine.loader.FlutterLoader;
 import io.flutter.plugin.common.BinaryMessenger;
+
 import org.jetbrains.annotations.NotNull;
 import org.json.JSONObject;
 
@@ -92,7 +99,9 @@ public class InstabugApi implements InstabugPigeon.InstabugHostApi {
 
     @NotNull
     @Override
-    public Boolean isBuilt() { return Instabug.isBuilt(); }
+    public Boolean isBuilt() {
+        return Instabug.isBuilt();
+    }
 
     @Override
     public void init(@NonNull String token, @NonNull List<String> invocationEvents, @NonNull String debugLogsLevel) {
@@ -441,4 +450,24 @@ public class InstabugApi implements InstabugPigeon.InstabugHostApi {
     public void willRedirectToStore() {
         Instabug.willRedirectToStore();
     }
+
+    public static void setScreenshotCaptor(ScreenshotCaptor screenshotCaptor,InternalCore internalCore) {
+        internalCore._setScreenshotCaptor(new com.instabug.library.screenshot.ScreenshotCaptor() {
+            @Override
+            public void capture(@NonNull ScreenshotRequest screenshotRequest) {
+                screenshotCaptor.capture(new ScreenshotCaptor.CapturingCallback() {
+                    @Override
+                    public void onCapturingFailure(Throwable throwable) {
+                        screenshotRequest.getListener().onCapturingFailure(throwable);
+                    }
+
+                    @Override
+                    public void onCapturingSuccess(Bitmap bitmap) {
+                        screenshotRequest.getListener().onCapturingSuccess(bitmap);
+                    }
+                });
+            }
+        });
+    }
+
 }