+++ packages/instabug_private_views/android/src/main/java/com/instabug/instabug_private_views/modules/InstabugPrivateView.java
@@ -2,31 +2,38 @@ package com.instabug.instabug_private_views.modules;
 
 import androidx.annotation.NonNull;
 
+import com.instabug.flutter.generated.InstabugLogPigeon;
+import com.instabug.flutter.modules.InstabugApi;
+import com.instabug.flutter.modules.InstabugLogApi;
+import com.instabug.flutter.util.privateViews.ScreenshotCaptor;
 import com.instabug.instabug_private_views.generated.InstabugPrivateViewPigeon;
 import com.instabug.library.internal.crossplatform.InternalCore;
-import com.instabug.library.screenshot.ScreenshotCaptor;
 import com.instabug.library.screenshot.instacapture.ScreenshotRequest;
 
 import io.flutter.plugin.common.BinaryMessenger;
 
 public class InstabugPrivateView implements InstabugPrivateViewPigeon.InstabugPrivateViewHostApi {
     PrivateViewManager privateViewManager;
-    private final InternalCore internalCore;
 
-    public InstabugPrivateView(BinaryMessenger messenger, PrivateViewManager privateViewManager, InternalCore internalCore) {
+    public static void init(BinaryMessenger messenger, PrivateViewManager privateViewManager) {
+        final InstabugPrivateView api = new InstabugPrivateView(messenger,privateViewManager);
+        InstabugPrivateViewPigeon.InstabugPrivateViewHostApi.setup(messenger, api);
+    }
+
+    public InstabugPrivateView(BinaryMessenger messenger, PrivateViewManager privateViewManager) {
         this.privateViewManager = privateViewManager;
-        this.internalCore = internalCore;
         InstabugPrivateViewPigeon.InstabugPrivateViewHostApi.setup(messenger, this);
 
     }
 
     @Override
     public void init() {
-        internalCore._setScreenshotCaptor(new ScreenshotCaptor() {
+        InstabugApi.setScreenshotCaptor(new ScreenshotCaptor() {
             @Override
-            public void capture(@NonNull ScreenshotRequest screenshotRequest) {
-                privateViewManager.mask(screenshotRequest.getListener());
+            public void capture(CapturingCallback listener) {
+                privateViewManager.mask(listener);
+
             }
-        });
+        },InternalCore.INSTANCE);
     }
 }