+++ packages/instabug_private_views/lib/src/repro_steps.dart
@@ -3,7 +3,7 @@ import 'package:instabug_private_views/src/private_views_manager.dart';
 
 // ignore: avoid_classes_with_only_static_members
 class ReproSteps {
-  static void enablePrivateViews() {
+  static void enableMaskingPrivateViews() {
     final api = InstabugPrivateViewHostApi();
     api.init();
     InstabugPrivateViewFlutterApi.setup(PrivateViewsManager.I);