+++ packages/instabug_private_views/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [UnReleased](https://github.com/Instabug/)
+
+### Added
+
+- Add support for masking private views during screen capturing ([#527](https://github.com/Instabug/Instabug-Flutter/pull/527)).